Prevent Unicode signature from messing up line parsing. Increase the size of PMID and REFMAN table columns in Citations table.

+++ abstrackr/model/__init__.py
@@ -97,8 +97,8 @@ class Citation(Base):
 
     # associates the citation with the project that owns it
     project_id = Column(types.Integer, ForeignKey('projects.id'))
-    pmid = Column(types.Integer)
-    refman = Column(types.Integer)
+    pmid = Column(types.BigInteger)
+    refman = Column(types.BigInteger)
 
     title = Column(types.Unicode(500))
     # length is based on back-of-envelope calculation
